+++ blocks/src/screener/css/style.scss
